+++ packages/server/src/modules/reports/resolvers/reports/yearly-ledger-report.resolver.ts
@@ -16,7 +16,7 @@ import {
 } from '@shared/gql-types';
 import { formatFinancialAmount } from '@shared/helpers';
 import type { TimelessDateString } from '@shared/types';
-import { sortEntityRecordsAndAddBalance } from '../../helpers/yearly-ledger-report.helper';
+import { sortEntityRecordsAndAddBalance } from '../../helpers/yearly-ledger-report.helper.js';
 
 export const yearlyLedgerReport: ResolverFn<
   ResolversTypes['YearlyLedgerReport'],
